+++ demo/android/app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
       android:name=".MainApplication"
       android:allowBackup="true"
       android:label="@string/app_name"
-      android:icon="@mipmap/ic_launcher"
+      android:icon="@mipmap/logo"
       android:theme="@style/AppTheme">
       <activity
         android:name=".MainActivity"
