+++ demo/package.json
@@ -11,6 +11,7 @@
     "react": "16.0.0-alpha.6",
     "react-native": "0.43.4",
     "react-native-cookies": "^3.2.0",
+    "react-native-divide": "^1.0.1",
     "react-native-simple-toast": "0.0.5",
     "react-navigation": "^1.0.0-beta.7"
   },
